requirements_file.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# HTTP Client
httpx==0.26.0

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# PDF Processing
pypdf2==3.0.1

# Image Processing
pillow==10.2.0
matplotlib==3.8.2
seaborn==0.13.1

# LLM APIs
google-generativeai==0.3.2
openai==1.10.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3